+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/case-editor/services/case-flag-state.service.ts
@@ -5,9 +5,11 @@ import { FormGroup } from '@angular/forms';
 export class CaseFlagStateService {
   public formGroup: FormGroup = new FormGroup({});
   public pageLocation: string;
+  public fieldStateToNavigate: number;
 
   public resetCache(pageLocation: string): void {
     this.formGroup = new FormGroup({});
+    this.fieldStateToNavigate = undefined;
     this.pageLocation = pageLocation;
   }
 }