+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/case-flag/read-case-flag-field.component.ts
@@ -6,7 +6,7 @@ import { CaseFlagStateService } from '../../case-editor/services/case-flag-state
 import { AbstractFieldReadComponent } from '../base-field/abstract-field-read.component';
 import { PaletteContext } from '../base-field/palette-context.enum';
 import { FlagDetailDisplay, FlagsWithFormGroupPath } from './domain';
-import { CaseFlagSummaryListDisplayMode } from './enums';
+import { CaseFlagStatus, CaseFlagSummaryListDisplayMode } from './enums';
 
 @Component({
   selector: 'ccd-read-case-flag-field',
@@ -20,7 +20,7 @@ export class ReadCaseFlagFieldComponent extends AbstractFieldReadComponent imple
   public paletteContext = PaletteContext;
   public flagForSummaryDisplay: FlagDetailDisplay;
   public summaryListDisplayMode: CaseFlagSummaryListDisplayMode;
-  public displayContextParameter = '';
+  public displayContextParameter: string;
 
   public pathToFlagsFormGroup: string;
   public readonly caseLevelCaseFlagsFieldId = 'caseFlags';
@@ -29,6 +29,7 @@ export class ReadCaseFlagFieldComponent extends AbstractFieldReadComponent imple
   private readonly createMode = '#ARGUMENT(CREATE)';
   private readonly createExternalMode = '#ARGUMENT(CREATE,EXTERNAL)';
   private readonly updateMode = '#ARGUMENT(UPDATE)';
+  private readonly updateExternalMode = '#ARGUMENT(UPDATE,EXTERNAL)';
 
   constructor(
     private readonly route: ActivatedRoute,
@@ -86,18 +87,23 @@ export class ReadCaseFlagFieldComponent extends AbstractFieldReadComponent imple
           this.summaryListDisplayMode = CaseFlagSummaryListDisplayMode.CREATE;
         // The FlagLauncher component holds a reference (selectedFlag), which gets set after the selection step of the
         // Manage Case Flags journey
-        } else if (flagLauncherComponent.caseField.display_context_parameter === this.updateMode &&
+        } else if ((flagLauncherComponent.caseField.display_context_parameter === this.updateMode ||
+          flagLauncherComponent.caseField.display_context_parameter === this.updateExternalMode) &&
           flagLauncherComponent.selectedFlag) {
             this.flagForSummaryDisplay =
               this.formGroup.get(flagLauncherControlName)['component'].selectedFlag.flagDetailDisplay;
           // TODO: not the best solution, the caseFlagStateService should have all the fields, then we can delete a lot of the transformations here
           // in Create Case Flag it already has all fields
-          this.flagForSummaryDisplay.flagDetail = {
+          const caseFlagFormGroupValue = this.caseFlagStateService.formGroup?.value;
+          if (caseFlagFormGroupValue) {
+            caseFlagFormGroupValue.status = CaseFlagStatus[caseFlagFormGroupValue.status];
+            this.flagForSummaryDisplay.flagDetail = {
               ...this.flagForSummaryDisplay.flagDetail,
-              ...this.caseFlagStateService.formGroup?.value
+              ...caseFlagFormGroupValue
             };
-            // Set the display mode for the "Review flag details" summary page
-            this.summaryListDisplayMode = CaseFlagSummaryListDisplayMode.MANAGE;
+          }
+          // Set the display mode for the "Review flag details" summary page
+          this.summaryListDisplayMode = CaseFlagSummaryListDisplayMode.MANAGE;
         }
       }
     }
@@ -124,11 +130,7 @@ export class ReadCaseFlagFieldComponent extends AbstractFieldReadComponent imple
   }
 
   public navigateBackToForm(fieldState: number): void {
-    this.router.navigate([`../${this.caseFlagStateService.pageLocation}`], {
-      relativeTo: this.route,
-      state: {
-        fieldState,
-      }
-    });
+    this.caseFlagStateService.fieldStateToNavigate = fieldState;
+    this.router.navigate([`../${this.caseFlagStateService.pageLocation}`], { relativeTo: this.route });
   }
 }